Updates June 13th
Fix for issue #4 SWS is no longer stackable
Fix for issue #10 Power station has fixed attach rules.

Shifted the lower plugs on the power station up a bit to reduce clipping
with the ground.
Made the collider on the plug smaller - it was larger than the model
which likely contributed to popping the power stations off the ground.
Added a bit more in the form of explanation to the central station and
plug part cfgs.